Moved the sidebar and added a AccountsTable to it
