+++ src/ui/server-administration/users/UserSidebar/index.ts
@@ -1,0 +1,1 @@
+export { UserSidebarContainer as UserSidebar } from './UserSidebarContainer';
